requirements/dev.txt
# aws-sam-cli
# aws-lambda-powertools==2.40.0
# コードで使用
# psycopg==3.1.19
# psycopg-binary==3.1.19
feedparser==6.0.11
asyncpg==0.29.0
aiohttp==3.9.5
aiogoogle==5.12.0
SQLAlchemy==2.0.31
alembic==1.13.2
python-dateutil==2.9.0.post0
pydantic==2.8.2
pytz==2024.1
sentry-sdk==2.10.0
# pytrends==4.9.2
beautifulsoup4==4.12.3
# テスト
pytest==8.2.2
pytest-mock==3.14.0
pytest-cov==5.0.0
pytest-asyncio==0.23.7
aioresponses==0.7.6